requirements.txt
Django==4.2
django-filter==24.3
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
drf-yasg==1.21.7
django-cors-headers==4.4.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
coverage==7.6.1
djangorestframework-simplejwt==5.3.1
setuptools==75.1.0
